Add more debugging info for future troubleshooting

+++ networkclient.cpp
@@ -78,7 +78,7 @@ void NetworkClient::downloadPrintFile(QJsonObject job) {
     qDebug() << "Print File ID: " << print_file_id;
 
     QUrl url = QUrl("http://10.0.2.2:3000/api/printmanager/v1_0/get_file/" + print_file_id ); //TODO: Move address and port to config file
-    qDebug() << "PDF URL: " << url.toString();
+    qDebug() << "DOWNLOADING PDF FROM URL: " << url.toString();
     QNetworkReply* reply = nam->get(QNetworkRequest(url));
     reply->setProperty("job", job);
 }
@@ -86,53 +86,70 @@ void NetworkClient::downloadPrintFile(QJsonObject job) {
 void NetworkClient::downloadPrintFileFinished(QNetworkReply *reply) {
     qDebug() << "NetworkClient::downloadPrintFileFinished";
 
-    QJsonObject job = reply->property("job").toJsonObject();
-    qDebug() << "JOB: " << QJsonDocument(job).toJson(QJsonDocument::Compact).toStdString().c_str();
-    qDebug() << "PRINT FILE ID: " << QString::number( job["job_id"].toInt() );
 
-    //QString jobId = reply->rawHeader("File-Id");
-    QString jobId = QString::number( job["job_id"].toInt() );
-    qDebug() << "FILE ID: " + jobId;
+    if (reply->error()) {
+        qDebug() << "ERROR!";
+        qDebug() << reply->errorString();
+    } else {
+        qDebug() << "1) " << reply->header(QNetworkRequest::ContentTypeHeader).toString();
+        qDebug() << "2) " << reply->header(QNetworkRequest::LastModifiedHeader).toDateTime().toString();
+        qDebug() << "3) " << reply->header(QNetworkRequest::ContentLengthHeader).toULongLong();
+        qDebug() << "4) " << reply->attribute(QNetworkRequest::HttpStatusCodeAttribute).toInt();
+        qDebug() << "5) " << reply->attribute(QNetworkRequest::HttpReasonPhraseAttribute).toString();
 
-    QNetworkAccessManager nam;
-    nam.get(QNetworkRequest(QUrl("http://10.0.2.2:3000/api/printmanager/v1_0/job/" + jobId + "/InProgress" )));
+        QJsonObject job = reply->property("job").toJsonObject();
+        qDebug() << "JOB: " << QJsonDocument(job).toJson(QJsonDocument::Compact).toStdString().c_str();
+        qDebug() << "PRINT FILE ID: " << QString::number( job["job_id"].toInt() );
 
-    QString tempDir = QDir::tempPath();
-    qDebug() << "Temp Dir: " << tempDir;
+        //QString jobId = reply->rawHeader("File-Id");
+        QString jobId = QString::number( job["job_id"].toInt() );
+        qDebug() << "FILE ID: " + jobId;
+        QString physicalPrinterName = job["physical_printer_name"].toString();
+        qDebug() << "PRINTER NAME: " << physicalPrinterName;
 
-    QString tempFile = tempDir + "/" + jobId + ".pdf";
-    QFile localFile( tempFile );
-    if (!localFile.open(QIODevice::WriteOnly))
-        return;
-    localFile.write(reply->readAll());
-    localFile.close();
+        QNetworkAccessManager nam;
+        nam.get(QNetworkRequest(QUrl("http://10.0.2.2:3000/api/printmanager/v1_0/job/" + jobId + "/InProgress" )));
 
-    QProcess sumatra;
-    qDebug() << "PRINTING TO START";
-    QString command = "C:\\SumatraPDF.exe -silent -print-to BrotherLaser1 " + tempFile;
-    qDebug() << "PRINT COMMAND: " << command;
-    sumatra.start(command);
-    sumatra.waitForStarted();
+        QString tempDir = QDir::tempPath();
+        qDebug() << "Temp Dir: " << tempDir;
 
-    qDebug() << "PRINTING STARTED";
+        QString tempFile = tempDir + "/" + jobId + ".pdf";
+        QFile localFile(tempFile);
+        if (localFile.open(QIODevice::WriteOnly)) {
+           localFile.write(reply->readAll());
+           localFile.flush();
+           localFile.close();
+        } else {
+            qDebug() << "Failed to open file for writing: " + tempFile;
+        }
 
-    sumatra.waitForFinished();
-    qDebug() << "PRINTING DONE!";
+        QProcess sumatra;
+        qDebug() << "PRINTING TO START";
+        QString command = "C:\\SumatraPDF.exe -silent -print-to " + physicalPrinterName + " " + tempFile;
+        qDebug() << "PRINT COMMAND: " << command;
+        sumatra.start(command);
+        sumatra.waitForStarted();
 
-    qDebug() << "EXIT STATUS: " << sumatra.exitCode();
-    if ( sumatra.exitStatus() == QProcess::NormalExit && sumatra.exitCode() == 0 ) {
+        qDebug() << "PRINTING STARTED";
+
+        sumatra.waitForFinished();
+        qDebug() << "PRINTING DONE!";
+
+        qDebug() << "EXIT STATUS: " << sumatra.exitCode();
+        if ( sumatra.exitStatus() == QProcess::NormalExit && sumatra.exitCode() == 0 ) {
         qDebug() << "PRINTING " << tempFile << " SUCEEDED!";
         nam.get(QNetworkRequest(QUrl("http://10.0.2.2:3000/api/printmanager/v1_0/job/" + jobId + "/Done" )));
-    } else {
+        } else {
         qDebug() << "PRINTING " << tempFile << " FAILED!";
         nam.get(QNetworkRequest(QUrl("http://10.0.2.2:3000/api/printmanager/v1_0/job/" + jobId + "/Error" )));
+        }
+
+        reply->deleteLater();
+
+        QFile file(tempFile);
+        file.remove();
+
+        qDebug() << "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX";
     }
-
-    reply->deleteLater();
-
-    QFile file(tempFile);
-    file.remove();
-
-    qDebug() << "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX";
 }
